+++ src/opm/parser/eclipse/EclipseState/Aquifer/Aquancon.cpp
@@ -17,7 +17,7 @@
   along with OPM.  If not, see <http://www.gnu.org/licenses/>.
  */
 #include <opm/parser/eclipse/EclipseState/Grid/FaceDir.hpp>
-#include <opm/parser/eclipse/EclipseState/Aquancon.hpp>
+#include <opm/parser/eclipse/EclipseState/Aquifer/Aquancon.hpp>
 #include <opm/common/utility/OpmInputError.hpp>
 #include <opm/common/OpmLog/OpmLog.hpp>
