+++ src/opm/parser/eclipse/EclipseState/Aquifer/AquiferConfig.cpp
@@ -18,7 +18,7 @@
 */
 
 
-#include <opm/parser/eclipse/EclipseState/AquiferConfig.hpp>
+#include <opm/parser/eclipse/EclipseState/Aquifer/AquiferConfig.hpp>
 #include <opm/parser/eclipse/EclipseState/Tables/TableManager.hpp>
 #include <opm/parser/eclipse/EclipseState/Grid/EclipseGrid.hpp>
