+++ src/opm/parser/eclipse/EclipseState/Aquifer/Aquifetp.cpp
@@ -20,7 +20,7 @@
 #include <utility>
 
 #include <opm/parser/eclipse/Parser/ParserKeywords/A.hpp>
-#include <opm/parser/eclipse/EclipseState/Aquifetp.hpp>
+#include <opm/parser/eclipse/EclipseState/Aquifer/Aquifetp.hpp>
 #include <opm/parser/eclipse/Deck/Deck.hpp>
 #include <opm/parser/eclipse/Deck/DeckRecord.hpp>
 #include <opm/common/utility/OpmInputError.hpp>
